refactor: add names of libraries entrypoints in macros
