--- types/LibraryType.hpp
@@ -1,16 +1,0 @@
-/*
-** EPITECH PROJECT, 2024
-** arcade-shared
-** File description:
-** LibraryType
-*/
-
-#pragma once
-
-namespace shared::types
-{
-  enum LibraryType{
-    GAME,
-    GRAPHIC,
-  };
-}
